% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{get_upstream}
\alias{get_upstream}
\title{Get upstream intervals}
\usage{
get_upstream(gr, width = 5000L)
}
\arguments{
\item{gr}{A GenomicRanges object.}

\item{width}{Width of the upstream interval. Default 5000.}
}
\value{
FIXME
}
\description{
Get upstream intervals from a GRanges object.
}
\examples{
gr <- data.frame(seqnames = "chr1",
                 strand = c("+", "-"),
                 start = c(42001, 67001),
                 width = c(1000, 1000),
                 gene = c("A", "B")) |>
  plyranges::as_granges()
gr
get_upstream(gr)
}
