% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{get_upstream_interval}
\alias{get_upstream_interval}
\title{Title}
\usage{
get_upstream_interval(gr, width = 5000L)
}
\arguments{
\item{gr}{A GenomicRanges object}

\item{width}{Width of the upstream interval}
}
\value{
fixme
}
\description{
Title
}
\examples{
gr <- data.frame(seqnames = "chr1",
                 strand = c("+", "-"),
                 start = c(42001, 67001),
                 width = c(1000, 1000),
                 gene = c("A", "B")) |>
  plyranges::as_granges()
gr
get_upstream_interval(gr)
}
