% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{get_upstream_snps}
\alias{get_upstream_snps}
\title{Get SNPs in upstream regions}
\usage{
get_upstream_snps(snps, txdb, level = "genes", ...)
}
\arguments{
\item{snps}{SNPs read in with \link{read_vcf}.}

\item{txdb}{A \link[GenomicFeatures:TxDb-class]{GenomicFeatures::TxDb} object with transcript annotations for the organism of interest. Must match the organism specified in the \code{bsgenome} argument of \link{read_vcf}.}

\item{level}{Currently only \code{"genes"} and \code{"transcripts"} are supported, which run \link[GenomicFeatures:transcripts]{GenomicFeatures::genes} and \link[GenomicFeatures:transcripts]{GenomicFeatures::transcripts}, respectively.}

\item{...}{Additional arguments passed to \link{get_upstream} (e.g., \verb{width=}).}
}
\value{
A GRanges object containing SNPs in regions upstream of intervals in the specified TxDb. See \link{get_upstream}.
}
\description{
Get SNPs upstream of gene regions. Input SNPs read in with \link{read_vcf} and an appropriate TxDb object.
}
\examples{
\dontrun{
library(BSgenome.Ggallus.UCSC.galGal6)
library(TxDb.Ggallus.UCSC.galGal6.refGene)
bs <- BSgenome.Ggallus.UCSC.galGal6
tx <- TxDb.Ggallus.UCSC.galGal6.refGene
vcf_file <- system.file("extdata", "galGal6-chr33.vcf.gz", package="tfboot", mustWork = TRUE)
snps <- read_vcf(vcf_file, BSgenome.Ggallus.UCSC.galGal6)
upstreamsnps <- get_upstream_snps(snps=snps, txdb=tx, level="genes")
upstreamsnps
}
}
