% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{mb_bootstats}
\alias{mb_bootstats}
\title{Bootstrap statistics}
\usage{
mb_bootstats(mbsmry, mbboot)
}
\arguments{
\item{mbsmry}{Results from running \link{mb_summarize} on motifbreakR results from your gene set of interest.}

\item{mbboot}{Results from running \link{mb_bootstrap} on a full background of all genes. Typically this should be performed once, with results read in from a file.}
}
\value{
A tibble with each metric from your bootstrap resampling (see \link{mb_summarize}), and p-value comparing the actual value of your gene set (\code{stat}) against the empirical null distribution (\code{bootdist}).
}
\description{
Get statistics (p-values) on your gene set's motifbreakR results compared to the bootstrapped empirical null distribution.
}
\examples{
data(vignettedata)
mbres <- vignettedata$mbres
mball <- vignettedata$mball
mbsmry <- mb_summarize(mbres)
mbsmry
set.seed(42)
mbboot <- mb_bootstrap(mball, ngenes=5, boots = 100)
mbboot
mb_bootstats(mbsmry, mbboot)
}
