% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{mb_bootstrap}
\alias{mb_bootstrap}
\title{Bootstrap motifbreakR results}
\usage{
mb_bootstrap(mbtibble, ngenes, boots = 100, key_col = "gene_id")
}
\arguments{
\item{mbtibble}{A tibble of motifbreakR results from \link{mb_to_tibble}.}

\item{ngenes}{The number of genes to sample with each bootstrap resample.}

\item{boots}{The number of bootstrap resamples.}

\item{key_col}{The name of the column used to key the txdb. Default \code{gene_id}. May be \code{transcript_id} or otherwise if you use a different value of \code{level} in \link{get_upstream_snps}.}
}
\value{
A list of two tibbles. See description and examples.
}
\description{
Bootstrap motifbreakR results. Takes motifbreakR results as a tibble (from
\link{mb_to_tibble}), draws \code{boots} random samples of \code{ngenes} genes, and returns
as a list (1) a wide tibble with results for each bootstrap, and (2) another
tibble with the distribution of each metric in a listcol. See examples.
}
\details{
Typically, you want to run this function on a full set of all genes to create
an empirical null distribution. You should run motifbreakR \emph{once} on all
genes, save this as an RData object, and read it in during bootstrap
resampling. See vignettes.
}
\examples{
data(vignettedata)
vignettedata$mball
mb_bootstrap(vignettedata$mball, ngenes=5, boots=5)
}
