% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{mb_summarize}
\alias{mb_summarize}
\title{Summarize motifbreakR results}
\usage{
mb_summarize(mbtibble, key_col = "gene_id")
}
\arguments{
\item{mbtibble}{motifbreakR results summarized with \link{mb_to_tibble}.}

\item{key_col}{The name of the column used to key the txdb. Default \code{gene_id}. May be \code{transcript_id} or otherwise if you use a different value of \code{level} in \link{get_upstream_snps}.}
}
\value{
A tibble. See description.
}
\description{
Summarizes motifbreakR results as a tibble from \link{mb_to_tibble}. See details.
}
\details{
Summarizes motifbreakR results. Returns a tibble with columns indicating:
\enumerate{
\item \code{nsnps}: The number of SNPs total
\item \code{nstrong}: The number of SNPs with a "strong" effect
\item \code{alleleDiffAbsMean} The mean of the absolute values of the \code{alleleDiff} scores
\item \code{alleleDiffAbsSum} The sum of the absolute values of the \code{alleleDiff} scores
\item \code{alleleEffectSizeAbsMean} The mean of the absolute values of the \code{alleleEffectSize} scores
\item \code{alleleEffectSizeAbsSum} The sum of the absolute values of the \code{alleleEffectSize} scores
}
}
\examples{
data(vignettedata)
vignettedata$mbres
mb_summarize(vignettedata$mbres)
}
