% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{parallel_motifbreakR}
\alias{parallel_motifbreakR}
\title{Title}
\usage{
parallel_motifbreakR(grl, cpus = NULL, ...)
}
\arguments{
\item{grl}{FIXME}

\item{cpus}{FIXME}

\item{...}{FIXME}
}
\value{
FIXME
}
\description{
Title
}
