% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{plot_bootstats}
\alias{plot_bootstats}
\title{Plot bootstrap distributions}
\usage{
plot_bootstats(bootstats)
}
\arguments{
\item{bootstats}{Output from running \link{mb_bootstats} on your results and a bootstrap resampling.}
}
\value{
A ggplot2 plot object. See description and examples.
}
\description{
Plot bootstrap distributions of motifbreakR results with your critical value highlighted by a vertical red line.
}
\examples{
data(vignettedata)
mbres <- vignettedata$mbres
mball <- vignettedata$mball
mbsmry <- mb_summarize(mbres)
set.seed(42)
mbboot <- mb_bootstrap(mball, ngenes=5, boots = 250)
bootstats <- mb_bootstats(mbsmry, mbboot)
plot_bootstats(bootstats)
}
