% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{read_vcf}
\alias{read_vcf}
\title{Read in SNPs from a VCF}
\usage{
read_vcf(file, bsgenome)
}
\arguments{
\item{file}{File path to a VCF file. See details.}

\item{bsgenome}{An object of class BSgenome for the species you are interrogating; see \link[BSgenome:available.genomes]{BSgenome::available.genomes} for a list of species.}
}
\value{
A GRanges object containing \code{SNP_id}, \code{REF}, and \code{ALT} columns.
}
\description{
Helper function to read in SNP data from a VCF file.
}
\details{
Note that the VCF \strong{must} be filtered to only contain variant sites (i.e.,
no \code{0/0}), or only homozygous alt sites if you choose (\code{0/1} or \code{1/1}). This
can be accomplished with bcftools:

\verb{# Filter to any variant sites:} \cr
\verb{bcftools view -i 'GT="alt"' ...} \cr
\verb{# Filter to homozygous alt sites:} \cr
\verb{bcftools view -i 'GT="AA"' ...}
}
\examples{
\dontrun{
library(BSgenome.Ggallus.UCSC.galGal6)
vcf_file <- system.file("extdata", "galGal6-chr33.vcf.gz", package="tfboot", mustWork = TRUE)
snps <- read_vcf(vcf_file, BSgenome.Ggallus.UCSC.galGal6)
snps
}
}
