% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{split_gr_by_id}
\alias{split_gr_by_id}
\title{Split GRanges by gene}
\usage{
split_gr_by_id(gr, split_col = "gene_id")
}
\arguments{
\item{gr}{A GRanges object returned by \link{get_upstream_snps}.}

\item{split_col}{The name of the column in \code{gr} to split by (default \code{gene_id}).}
}
\value{
FIXME
}
\description{
Splits a GRanges object into a GRangesList by a column (typically \code{gene_id})
}
\examples{
gr <- data.frame(seqnames=rep(c("chr1", "chr2", "chr1", "chr3"), c(1, 3, 2, 4)),
           start=1:10,
           width=1,
           gene_id = rep(c("gene1", "gene2", "gene3", "gene4"), c(4, 2, 1, 3))) |>
  plyranges::as_granges()
gr
split_gr_by_id(gr, split_col="gene_id")
}
