% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{summarize_mb}
\alias{summarize_mb}
\title{Summarize motifbreakR results}
\usage{
summarize_mb(mbtibble)
}
\arguments{
\item{mbtibble}{motifbreakR results summarized with \link{mb_to_tibble}.}
}
\value{
A tibble. See description.
}
\description{
Summarizes motifbreakR results as a tibble from \link{mb_to_tibble}. See details.
}
\details{
Summarizes motifbreakR results. Returns a tibble with columns indicating:
\enumerate{
\item \code{nsnps}: The number of SNPs total
\item \code{nstrong}: The number of SNPs with a "strong" effect
\item \code{alleleDiffAbsMean} The mean of the absolute values of the \code{alleleDiff} scores
\item \code{alleleDiffAbsSum} The sum of the absolute values of the \code{alleleDiff} scores
\item \code{alleleEffectSizeAbsMean} The mean of the absolute values of the \code{alleleEffectSize} scores
\item \code{alleleEffectSizeAbsSum} The sum of the absolute values of the \code{alleleEffectSize} scores
}
}
