% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot-package.R
\docType{package}
\name{tfboot-package}
\alias{tfboot}
\alias{tfboot-package}
\title{tfboot: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Stephen Turner \email{turner@colossal.com} (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})

Other contributors:
\itemize{
  \item Colossal Biosciences [copyright holder]
}

}
\keyword{internal}
