% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignettedata}
\alias{vignettedata}
\title{Vignette data}
\format{
An object of class \code{list} of length 2.
}
\usage{
vignettedata
}
\description{
Pre-cooked data used by the vignette. This data is precomputed
to speed vignette compilation time and is used throughout the examples.
Contains a list of two objects:
\itemize{
\item \code{mbres}: Results from running motifbreakR on five randomly selected genes, then run through \link{mb_to_tibble}.
\item \code{mball}: Results from running motifbreakR on \emph{all} genes, then run through \link{mb_to_tibble}.
}
}
\keyword{datasets}
