% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{get_upstream_snps}
\alias{get_upstream_snps}
\title{Get upstream SNPs}
\usage{
get_upstream_snps(snps, txdb, level = "genes", ...)
}
\arguments{
\item{snps}{FIXME}

\item{txdb}{FIXME}

\item{level}{FIXME}

\item{...}{Additional arguments passed to \link{get_upstream}.}
}
\value{
FIXME
}
\description{
Get SNPs upstream of gene regions. Input SNPs read in with \link{read_vcf} and an appropriate TxDb object.
}
\examples{
\dontrun{
library(BSgenome.Cfamiliaris.UCSC.canFam3)
library(TxDb.Cfamiliaris.UCSC.canFam3.ncbiRefSeq)
bs <- BSgenome.Cfamiliaris.UCSC.canFam3
tx <- TxDb.Cfamiliaris.UCSC.canFam3.ncbiRefSeq
vcf_file <- system.file("extdata", "rosie38.vcf.gz", package="tfboot", mustWork = TRUE)
snps <- read_vcf(vcf_file, BSgenome.Cfamiliaris.UCSC.canFam3)
snps
upstreamsnps <- get_upstream_snps(snps=snps, txdb=tx, level="genes")
upstreamsnps
}
}
