% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{mb_to_tibble}
\alias{mb_to_tibble}
\title{motifbreakR results to tibble}
\usage{
mb_to_tibble(mb, key_col = "gene_id", ...)
}
\arguments{
\item{mb}{motifbreakR results}

\item{key_col}{The name of the column used to key the txdb. Default \code{gene_id}.}
}
\description{
Make a compact tibble with only relevant columns from motifbreakR results
}
