% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot.R
\name{mb_to_tibble}
\alias{mb_to_tibble}
\title{motifbreakR results to tibble}
\usage{
mb_to_tibble(mb, key_col = "gene_id", ...)
}
\arguments{
\item{mb}{motifbreakR results GRanges object.}

\item{key_col}{The name of the column used to key the txdb. Default \code{gene_id}. May be \code{transcript_id} or otherwise if you use a different value of \code{level} in \link{get_upstream_snps}.}
}
\value{
A tibble containing the key column (usually \code{gene_id}), and a select number of other columns needed for downstream statistical analysis.
}
\description{
Make a compact tibble with only select columns from motifbreakR results GRanges objects.
}
