% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfboot-package.R
\docType{package}
\name{tfboot-package}
\alias{tfboot}
\alias{tfboot-package}
\title{tfboot: Bootstrapping and statistical analysis for TFBS-disrupting SNPs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Bootstrap motifbreakR results for statistical analysis on TFBS-disrupting SNPs in upstream regions of a set of genes of interest.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/colossal-compsci/tfboot}
  \item Report bugs at \url{https://github.com/colossal-compsci/tfboot/issues}
}

}
\author{
\strong{Maintainer}: Stephen Turner \email{turner@colossal.com} (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})

Other contributors:
\itemize{
  \item Colossal Biosciences [copyright holder]
}

}
\keyword{internal}
